move all our file formatters to the format dir
