+++ lib/python/qmk/cli/__init__.py
@@ -38,7 +38,10 @@ subcommands = [
     'qmk.cli.doctor',
     'qmk.cli.fileformat',
     'qmk.cli.flash',
+    'qmk.cli.format.c',
     'qmk.cli.format.json',
+    'qmk.cli.format.python',
+    'qmk.cli.format.text',
     'qmk.cli.generate.api',
     'qmk.cli.generate.config_h',
     'qmk.cli.generate.dfu_header',
